Weather check to ensure it's good or not to go outside using if else statements
Created index.html & script.js files for weather to check on which temperature what to do using if else statements.

A variation of the if statement is an if statement with multiple else if blocks.
This can be more efficient in certain situations because you are always only going to execute one or zero blocks. If you have many if else statements underneath one another, they are going to be evaluated and possibly executed even though one of
the ones above already had a condition evaluate to true and proceeded to execute the associated code block.